exracts function to its own module
