+++ 01/createStatementData.ts
@@ -1,0 +1,32 @@
+import { amountFor } from './amountFor.ts';
+import { playFor } from './playFor.ts';
+import totalAmount from './totalAmount.ts';
+import { totalVolumeCredits } from './totalVolumeCredits.ts';
+import type { Invoice, PlayPerformance, Plays } from './types.ts';
+import { volumeCreditsFor } from './volumeCreditsFor.ts';
+
+export function createStatementData(invoice: Invoice, plays: Plays) {
+  const statementData: Invoice = {
+    customer: '',
+    performances: [],
+    totalAmount: 0,
+    totalVolumeCredits: 0,
+  };
+
+  statementData.customer = invoice.customer;
+  statementData.performances = invoice.performances.map(enrichPerformance);
+  statementData.totalAmount = totalAmount(statementData);
+  statementData.totalVolumeCredits = totalVolumeCredits(statementData);
+
+  function enrichPerformance(aPerformance: PlayPerformance) {
+    const result: PlayPerformance = Object.assign({}, aPerformance);
+
+    result.play = playFor(result);
+    result.amount = amountFor(result);
+    result.volumeCredits = volumeCreditsFor(result);
+
+    return result;
+  }
+
+  return statementData;
+}
