+++ 01/statement.ts
@@ -1,39 +1,9 @@
-import { amountFor } from './amountFor.ts';
-import { playFor } from './playFor.ts';
-import totalAmount from './totalAmount.ts';
-import { totalVolumeCredits } from './totalVolumeCredits.ts';
-import { Invoice, Plays, PlayPerformance } from './types.ts';
+import { createStatementData } from './createStatementData.ts';
+import { Invoice, Plays } from './types.ts';
 import { usd } from './usd.ts';
-import { volumeCreditsFor } from './volumeCreditsFor.ts';
 
 export function statement(invoice: Invoice, plays: Plays) {
   return renderPlaintText(createStatementData(invoice, plays));
-}
-
-function createStatementData(invoice: Invoice, plays: Plays) {
-  const statementData: Invoice = {
-    customer: '',
-    performances: [],
-    totalAmount: 0,
-    totalVolumeCredits: 0,
-  };
-
-  statementData.customer = invoice.customer;
-  statementData.performances = invoice.performances.map(enrichPerformance);
-  statementData.totalAmount = totalAmount(statementData);
-  statementData.totalVolumeCredits = totalVolumeCredits(statementData);
-
-  function enrichPerformance(aPerformance: PlayPerformance) {
-    const result: PlayPerformance = Object.assign({}, aPerformance);
-
-    result.play = playFor(result);
-    result.amount = amountFor(result);
-    result.volumeCredits = volumeCreditsFor(result);
-
-    return result;
-  }
-
-  return statementData;
 }
 
 function renderPlaintText(data: Invoice) {
